Refactor: useUser 함수로 삭제한 세 함수를 통합
