--- src/hooks/user/useAuth.tsx
@@ -1,16 +1,0 @@
-import { useEffect, useState } from "react";
-import { getAuth, onAuthStateChanged, User } from "firebase/auth";
-
-const useAuth = () => {
-  const [user, setUser] = useState<User | null>(null);
-  const auth = getAuth();
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, setUser);
-    return () => unsubscribe();
-  }, [auth]);
-
-  return user;
-};
-
-export default useAuth;
