+++ src/hooks/user/useUser.ts
@@ -1,0 +1,36 @@
+import { useEffect, useState } from "react";
+import { getAuth, onAuthStateChanged, User } from "firebase/auth";
+import { getFirestore, doc, getDoc } from "firebase/firestore";
+
+const useUser = () => {
+  const [user, setUser] = useState<User | null>(null);
+  const [username, setUsername] = useState<string | null>(null);
+  const [profileImage, setProfileImage] = useState<string | null>(null);
+
+  const auth = getAuth();
+  const db = getFirestore();
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, setUser);
+    return () => unsubscribe();
+  }, [auth]);
+
+  useEffect(() => {
+    if (user?.uid) {
+      const fetchUserData = async () => {
+        const userDoc = doc(db, "users", user.uid);
+        const userSnap = await getDoc(userDoc);
+        if (userSnap.exists()) {
+          setUsername(userSnap.data().username);
+          setProfileImage(userSnap.data().profileImage);
+        }
+      };
+
+      fetchUserData();
+    }
+  }, [db, user]);
+
+  return { user, username, profileImage };
+};
+
+export default useUser;
